requirements.txt
pytesseract
easyocr
opencv-python
numpy
pandas
python-Levenshtein
matplotlib
Pillow
paddleocr
transformers
python-doctr[torch]
torch
torchvision
nltk
rouge-score
psutil
albumentations
